Create Word Of The Day Powershell Script
This is a work in progress. Using the Merriam-Webster dictionary API, I intend for this Powershell script to retrieve a "word of the day" and display it to the user.

Note!!!!!!: You will need to sign up for a free API key from Merriam-Webster to use this script which is annoying

To run this script once a day, you can use Windows Task Scheduler to schedule it to run at a specific time each day. 
(

Here are some directions for myself for the future I found online for how to do that:

1. Open Task Scheduler by searching for it in the Start menu.
2. Click "Create Basic Task" in the Actions pane.
3. Give your task a name and description, then click Next.
4. Set the trigger to "Daily" and choose the time you want the script to run.
5. Set the action to "Start a program" and enter the path to PowerShell.exe (probably going to be C:\Windows\System32\WindowsPowerShell\v1.0\PowerShell.exe).
6. In the "Add arguments" field, enter the path to your script file (e.g. C:\Scripts\WordOfTheDay.ps1).
7. Click Finish to save your task.
Now, your script will run automatically once a day at the scheduled time, and display the "word of the day" to the user.
)